Added a DeliverGear command and an OrientRobot command.
Added a DeliverGear command for auto mode that incorporates the gyro and the tracker (most likely not successfully), but does not yet incorporate
vision.
Also added an OrientRobot command that has the gyro (probably not
succesfully). It does not yet rotate the robot to a particular
orientation, which would be super useful to line up on pegs, goals,
hoppers, etc. in auto and teleop.